+++ src/helpers/excelExportHelper.ts
@@ -1,12 +1,5 @@
-import { ToolButton } from '@serenity-is/corelib'
 import { deepClone, postToService } from '@serenity-is/corelib/q'
 import { IdevsExportOptions, IdevsExportRequest } from '../globals'
-
-export type ExcelExportOptions = IdevsExportOptions & {
-  title?: string
-  hint?: string
-  separator?: boolean
-}
 
 export function onExportExcel(options: IdevsExportOptions): void {
   if (!options.onViewSubmit()) {
@@ -32,14 +25,4 @@ export function onExportExcel(options: IdevsExportOptions): void {
   request.viewName = options.viewName
 
   postToService({ service: options.service, request: request, target: '_blank' })
-}
-
-export function createExcelToolButton(options: ExcelExportOptions): ToolButton {
-  return {
-    hint: options.hint || 'Excel',
-    title: options.title || '',
-    cssClass: 'export-xlsx-button',
-    onClick: onExportExcel(options),
-    separator: options.separator,
-  }
 }